nogicos/requirements.txt
# NogicOS Dependencies

# Browser Automation
playwright>=1.40.0

# LLM Clients
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0

# Async Support
aioconsole>=0.6.0
nest-asyncio>=1.5.0

# Data Validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0
typer>=0.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

